+++ serve.py
@@ -50,20 +50,6 @@ def main(
         None,
         help="The faiss index to use for serving"
     ),
-    query_blocklist: Path = typer.Option(
-        None,
-        '--query-blocklist',
-        '--query-blacklist',
-        exists=True,
-        dir_okay=False,
-        file_okay=True,
-        readable=True,
-        help=(
-            "A text file containing a list of words/phrases (each separated by a line break) "
-            "that users should be blocked from searching. When the user enters a query that matches "
-            "one of the terms in the blocklist, an error message will be returned"
-        ),
-    )
 ):
     # ensure that the frontend assets are built
     if not Path(theme_asset_dir / 'index.html').exists():
@@ -77,7 +63,6 @@ def main(
         project_dir,
         theme_asset_dir,
         index_type.value if index_type else None,
-        query_blocklist
     )